requirements.txt
streamlit>=1.28.0
firebase-admin>=6.2.0
stripe>=5.5.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.1.1